Split meta mutations tests
